backend/core/requirements.txt
coreapi==2.3.3
Django==4.1.4
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.4
grpcio==1.51.1
grpcio-tools==1.51.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
pylint==2.12.2
pytest-django==4.5.2
python-decouple==3.6
requests==2.28.1
service-identity==21.1.0
simplejson==3.18.0
swagger-spec-validator==3.0.3
urllib3==1.26.13
virtualenv==20.16.6
yarg==0.1.9
